requirements.txt
google-generativeai
markdown2
pypandoc
PyPDF2
docx